FIX&Refactor(backend) aggiunta commenti e spostamento package per la foto profilo
