+++ bridge-backend/src/main/java/com/project/bridgebackend/fotoProfilo/FotoProfilo.java
@@ -1,4 +1,4 @@
-package com.project.bridgebackend.registrazione.fotoProfilo;
+package com.project.bridgebackend.fotoProfilo;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
