+++ bridge-backend/src/main/java/com/project/bridgebackend/fotoProfilo/FotoProfiloService.java
@@ -1,9 +1,8 @@
-package com.project.bridgebackend.registrazione.fotoProfilo;
+package com.project.bridgebackend.fotoProfilo;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
